Update service schema, remove feature gates from test.
